requirements.txt
path.py==10.1
docopt==0.6.2
-e git+https://github.com/OpenPathView/DirectoryManagerClient@master#egg=opv_directorymanagerclient
-e git+https://github.com/OpenPathView/OPV_DBRest-client@master#egg=opv_api_client
Pillow==4.0.0
